use global to store Br exit propagation;
clean up warnings;

+++ crates/codegen-tritonvm/src/ty.rs
@@ -10,7 +10,7 @@ pub fn felt_i32(v: i32) -> BFieldElement {
 
 pub fn felt_i64(v: i64) -> BFieldElement {
     if v < 0 {
-        BFieldElement::new(BFieldElement::P - (v.unsigned_abs() as u64))
+        BFieldElement::new(BFieldElement::P - v.unsigned_abs())
     } else {
         BFieldElement::new(v as u64)
     }
